Extract welcome page utilities to common utilities
